% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcVDYPMSYTyieldcurves-package.R
\docType{package}
\name{bcVDYPMSYTyieldcurves-package}
\alias{bcVDYPMSYTyieldcurves-package}
\title{bcVDYPMSYTyieldcurves: Process BC Forest Inventory Data and Generate Yield Curves via VDYP7}
\description{
Tools for processing British Columbia's Vegetation Resources Inventory (VRI) data and generating forest yield curves. The package provides functions to subset VRI data, match polygons to existing managed stand yield curves, identify records requiring VDYP7 processing, and execute the complete VDYP7 workflow including parameter file creation, execution, and post-processing of results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forest-carbon-tools-dfs/bcVDYPMSYTyieldcurves}
  \item Report bugs at \url{https://github.com/forest-carbon-tools-dfs/bcVDYPMSYTyieldcurves/issues}
}

}
\author{
\strong{Maintainer}: Derek Sattler \email{derek.sattler@nrcan-rncan.gc.ca}

Authors:
\itemize{
  \item Derek Sattler \email{derek.sattler@nrcan-rncan.gc.ca}
}

}
\keyword{internal}
