% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vdyp_parameter_file.R
\name{create_vdyp_parameter_file}
\alias{create_vdyp_parameter_file}
\title{Create VDYP7 Parameter File}
\usage{
create_vdyp_parameter_file(
  vdyp_dir,
  input_dir,
  output_dir,
  poly_file,
  layer_file,
  output_prefix = "VDYP7_OUTPUT",
  param_file_name = "parms.txt",
  utilization_cu = 12.5,
  age_start = 0,
  age_end = 250,
  age_increment = 10
)
}
\arguments{
\item{vdyp_dir}{Character. Path to the VDYP7 installation directory}

\item{input_dir}{Character. Path to directory containing input files}

\item{output_dir}{Character. Path to directory for output files}

\item{poly_file}{Character. Name of the polygon input CSV file}

\item{layer_file}{Character. Name of the layer input CSV file}

\item{output_prefix}{Character. Prefix for output file names. Default is "VDYP7_OUTPUT"}

\item{param_file_name}{Character. Name for the parameter file. Default is "parms.txt"}

\item{utilization_cu}{Numeric. Close utilization diameter in cm. Default is 12.5 (others: 4.0, 7.5, 17.5)}

\item{age_start}{Numeric. Starting age for projections. Default is 0}

\item{age_end}{Numeric. Ending age for projections. Default is 250}

\item{age_increment}{Numeric. Age increment for projections. Default is 10}
}
\value{
Character. Path to the created parameter file
}
\description{
Creates a parameter file (.txt) needed to run VDYP7console with specified
input files, output files and configuration options.
}
\details{
This function creates a parameter file that VDYP7console uses to
determine input files, output locations, and processing options. The parameter
file includes settings for:
- Input/output file locations
- Utilization standards by species group
- Age range and increment for projections
- Output format options

The function automatically creates the input and output directories if they
don't exist.

For help on VDYP7, go to:
https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/growth-and-yield-modelling/variable-density-yield-projection-vdyp/publications-and-support
}
\examples{
\dontrun{
# Create parameter file for VDYP7 run
param_file <- create_vdyp_parameter_file(
  vdyp_dir = "C:/VDYP7",
  input_dir = "C:/VDYP7/Input_testarea",
  output_dir = "C:/VDYP7/Output_testarea",
  poly_file = "vdyp_input_poly_test_area.csv",
  layer_file = "vdyp_input_layer_test_area.csv",
  param_file_name = "parms_testarea.txt"
)
}

}
