% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vri_for_vdyp.R
\name{get_vri_for_vdyp}
\alias{get_vri_for_vdyp}
\title{Get VDYP Input Files for Specific VRI Records}
\usage{
get_vri_for_vdyp(
  vri_file,
  msyt_reference_file,
  vdyp_input_dir,
  output_dir,
  layer_csv_file_name,
  poly_csv_file_name
)
}
\arguments{
\item{vri_file}{The full path to the VRI subset CSV file containing feature IDs.}

\item{msyt_reference_file}{The full path to the MSYT VDYP reference CSV file.}

\item{vdyp_input_dir}{The directory containing the raw VDYP input `poly` and `layer`
CSV files (e.g., `VEG_COMP_VDYP7_INPUT_POLY.csv` and `VEG_COMP_VDYP7_INPUT_LAYER.csv`).}

\item{output_dir}{The directory where the final, filtered poly and layer files will be saved.}

\item{layer_csv_file_name}{The name of the csv file to be written with vdyp layer info.}

\item{poly_csv_file_name}{The name of the csv file to be written with vdyp poly info.}
}
\value{
A list containing two data.tables: `poly_dt_final` and `layer_dt_final`.
  These data.tables are saved as CSV files in the specified `output_dir`.
}
\description{
This function identifies and extracts the VDYP input poly and layer records
that correspond to the VRI records within the area of interest (aoi) whose
current yield is based on VDYP projections. It performs an intersection of
feature IDs from the VRI data, and the raw VDYP poly and layer files.
}
\examples{
# Assuming you have the necessary files in a 'data' folder
# vri_file <- "data/vri_subset.csv"
# msyt_reference_file <- "data/msyt_2024_reference_VDYP.csv"
# vdyp_input_dir <- "data/vdyp_input"
# output_dir <- "output"
# layer_csv_file_name <- "vdyp_input_layer_aoi.csv"
# poly_csv_file_name <- "vdyp_input_poly_aoi.csv"

# vdyp_data <- get_vdyp_inputs(
#   vri_file = vri_file,
#   msyt_reference_file = msyt_reference_file,
#   vdyp_input_dir = vdyp_input_dir,
#   output_dir = output_dir,
#   layer_csv_file_name = layer_csv_file_name,
#   poly_csv_file_name = poly_csv_file_name
# )

# View the head of the filtered poly data
# head(vdyp_data$poly_dt_final)

}
