% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vydp_output.R
\name{process_vdyp_output}
\alias{process_vdyp_output}
\title{Process VDYP7 Output and Calculate Softwood/Hardwood Volume}
\usage{
process_vdyp_output(
  output_file,
  input_poly_file,
  output_dir,
  vdyp_postpro_file
)
}
\arguments{
\item{output_file}{The full path to the VDYP7 output CSV file
(e.g., "VDYP7_OUTPUT_YLDTBL_parms_aoi.csv").}

\item{input_poly_file}{The full path to the VDYP7 input polygon CSV file.}

\item{output_dir}{The directory where the final processed data
will be saved.}

\item{vdyp_postpro_file}{File name given to csv containing post-processed vdyp output.}
}
\value{
A data.table containing the processed VDYP output, with new columns
  for conifer and deciduous volumes. The function also saves the
  final data table as a CSV file.
}
\description{
This function takes the raw output from a VDYP7 run, joins it with the input
polygon and layer data, and calculates the conifer and deciduous volume components
based on species codes. It also saves the final, processed data table.
}
\examples{
\dontrun{
# Assuming the required files are available
# output_file <- "C:/VDYP7/Output/VDYP7_OUTPUT_YLDTBL_parms_aoi.csv"
# input_poly_file <- "C:/VDYP7/Input/poly_csv_file_name.csv"
# output_dir <- "C:/VDYP7/Output"
# vdyp_postpro_file <- "vdyp_postpro_file.csv"

# processed_data <- process_vdyp_output(
#   output_file = output_file,
#   input_poly_file = input_poly_file,
#   output_dir = output_dir
# )

# head(processed_data)
}
}
