% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vdyp7.R
\name{run_vdyp7}
\alias{run_vdyp7}
\title{Execute VDYP7console and Read Output}
\usage{
run_vdyp7(vdyp_exe, parms_file, output_dir, expected_output_file = NULL)
}
\arguments{
\item{vdyp_exe}{The full path to the VDYP7console executable.}

\item{parms_file}{The full path to the VDYP7 parameter file.}

\item{output_dir}{The directory where the VDYP7 output file is expected to be saved.}

\item{expected_output_file}{Character or NULL. Specific output CSV filename to read.
If NULL (default), the function will search for any CSV file in the output directory
and use the first one found (careful!). If specified, the function will look
for this exact filename in the output directory.}
}
\value{
A tibble containing the VDYP7 yield data if the output file is
  successfully generated and found. If the file is not found, a message
  is printed to the console and NULL is returned.
}
\description{
This function executes the external VDYP7console program using specified
executable and parameter files, then reads the generated yield table output.
VDYP7 and vdyp7console are available on the BC govn't website. When installing
VDYP7, it is recommended that it be placed as a sub-folder on the C drive
(e.g., C:/VDYP7).
}
\examples{
\dontrun{
# Assuming you have the VDYP executable and parameter files set up
# on your system and have prepared the input files.
# vdyp_path <- "C:/VDYP7/vdyp7console.exe"
# parms_file <- "C:/VDYP7/Input/my_parms.txt"
# output_dir <- "C:/VDYP7/Output"

# Run the model and read the output
# The function returns the data frame, so we can save it to a variable
# for later use.
yield_table <- run_vdyp7(
  vdyp_exe = vdyp_path,
  parms_file = parms_file,
  output_dir = output_dir
)

# If the run was successful, print the head of the output
if (!is.null(yield_table)) {
  head(yield_table)
}
}
}
