% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vri_to_csv.R
\name{write_vri_to_csv}
\alias{write_vri_to_csv}
\title{Write a VRI (or sf) Data Frame to a CSV File}
\usage{
write_vri_to_csv(sf_df, file_path, ...)
}
\arguments{
\item{sf_df}{An `sf` data frame (e.g., VRI data) to be written.}

\item{file_path}{A character string specifying the file path for the output CSV.}

\item{...}{Other arguments passed on to `data.table::fwrite`.}
}
\value{
Returns `invisible(NULL)`. The function's primary effect is creating
  a CSV file at the specified file path.
}
\description{
This function writes a spatial data frame (`sf` object) to a CSV file.
It automatically removes the geometry column and any other complex
list-type columns that are not supported by `data.table::fwrite`. The
resulting CSV file is used by the function 'get_vri_for_vdyp'.
}
\examples{
\dontrun{
# Assuming 'vri_subset' is an sf object created from your
# create_vri_subset() function

# Write the data frame to a CSV file.
# The geometry column will be automatically removed.
write_vri_to_csv(vri_subset, file.path(tempdir(), "vri_subset.csv"))
}
}
